requirements.txt
# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Dataset and Data Processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Retrieval Libraries
pyserini>=0.21.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Hallucination Detection
ragas>=0.1.0

# Evaluation Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
tqdm>=4.66.0
jsonlines>=4.0.0
python-dotenv>=1.0.0
wikipedia>=1.4.0

# Text Processing
nltk>=3.8.0
spacy>=3.6.0

# API Access (optional)
openai>=1.0.0

# Jupyter for exploration (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
